zezenia/requirements.txt
beautifulsoup4==4.11.2
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-bigquery==3.6.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
greenlet==2.0.2
grpcio==1.51.3
grpcio-status==1.51.3
halo==0.0.31
idna==3.4
log-symbols==0.0.14
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
platformdirs==3.1.1
proto-plus==1.22.2
protobuf==4.22.1
psycopg2==2.9.5
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
simple-term-menu==1.6.1
six==1.16.0
soupsieve==2.4
spinners==0.0.24
SQLAlchemy==2.0.5.post1
termcolor==2.2.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
